ISSUE-057 - Refactor: Switched from buildRow() to a dedicated component named TimingTowerRow. This is both simply a better way of going about it and will allow me to use hooks to track position changes and allow for new non-CSS animations.
