+++ src/features/broadcast/graphics/broadcastGraphicsSlice.ts
@@ -3,7 +3,7 @@ import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import {
   BGChyrons,
   BGStatusIndicatorModes,
-  BGTimingTowerDisplayModes,
+  BGTimingTowerDisplayMode,
   BGTimingTowerFocusedCarsMode,
   BGTimingTowerSplitsMode,
   OpenState,
@@ -20,7 +20,7 @@ const initialState: RootState["broadcastGraphics"] = {
     },
     timingTower: {
       open: 1,
-      displayMode: BGTimingTowerDisplayModes.LeftAndRight,
+      displayMode: BGTimingTowerDisplayMode.LeftAndRight,
       splitsMode: BGTimingTowerSplitsMode.Leader,
       focusedCars: [],
       focusedCarsMode: BGTimingTowerFocusedCarsMode.None,
